+++ tests/benchmark/sha512.cpp
@@ -9,7 +9,7 @@ TEST_CASE("sha512 measurements") {
 	std::array<std::byte, 1024> input{};
 
 	for (int i = 0; i != (int)input.size(); ++i) {
-		input[i] = static_cast<std::byte>(i);
+		input[static_cast<size_t>(i)] = static_cast<std::byte>(i);
 	}
 
 	BENCHMARK("16 byte input") {
@@ -65,7 +65,7 @@ TEST_CASE("openssl sha512 measurements") {
 	std::array<std::byte, 1024> input{};
 
 	for (int i = 0; i != (int)input.size(); ++i) {
-		input[i] = static_cast<std::byte>(i);
+		input[static_cast<size_t>(i)] = static_cast<std::byte>(i);
 	}
 
 	BENCHMARK("16 byte input") {
